- moved wrapper to core - get_event uses item factory
