+++ HABApp/openhab/events/item_events.py
@@ -60,7 +60,7 @@ class ItemCommandEvent(OpenhabEvent):
 
 
 class ItemAddedEvent(OpenhabEvent):
-    def __init__(self, name: str = '', type: str = None):
+    def __init__(self, name: str, type: str):
         super().__init__()
 
         self.name: str = name
@@ -78,7 +78,7 @@ class ItemAddedEvent(OpenhabEvent):
 
 
 class ItemUpdatedEvent(OpenhabEvent):
-    def __init__(self, name: str = '', type: str = None):
+    def __init__(self, name: str, type: str):
         super().__init__()
 
         self.name: str = name
