+++ HABApp/openhab/oh_connection.py
@@ -6,7 +6,7 @@ import HABApp
 import HABApp.core
 import HABApp.openhab.events
 from HABApp.openhab.map_events import get_event
-from HABApp.util import log_exception, ignore_exception
+from HABApp.core.wrapper import log_exception, ignore_exception
 from .http_connection import HttpConnection, HttpConnectionEventHandler
 from .oh_interface import get_openhab_interface
 from ..config import Openhab as OpenhabConfig
