+++ HABApp/util/__init__.py
@@ -1,5 +1,3 @@
-from .wrapper import log_exception, ignore_exception, ExceptionToHABApp
-
 from .timeframe import TimeFrame
 
 from .counter_item import CounterItem
